add a filter to the solr query
The field clustering that was used before was unique for our clusters, but the field analysis is not and retrieves other stuff from solr in addition to our clusters. I have added a quick fix for that which filters the query with everything that contains the word cluster which gives no errors and if we want to we can add some unique identifiers in the solr schema later.